Resolve audio provider review threads
