+++ src/core/providers/openai/client.rs
@@ -8,6 +8,10 @@ use std::collections::HashMap;
 use std::pin::Pin;
 use std::sync::Arc;
 
+use crate::core::audio::types::{
+    SpeechRequest, SpeechResponse, TranscriptionRequest, TranscriptionResponse, TranslationRequest,
+    TranslationResponse,
+};
 use crate::core::providers::base::{
     GlobalPoolManager, HeaderPair, HttpMethod, apply_headers, header, header_owned,
     read_streaming_error_body, send_streaming_request, streaming_unbounded_client,
@@ -27,7 +31,7 @@ use crate::core::types::{
 
 use super::{
     config::OpenAIConfig,
-    models::{OpenAIModelFeature, OpenAIModelRegistry, OpenAIUseCase, get_openai_registry},
+    models::{OpenAIModelRegistry, get_openai_registry},
 };
 use crate::core::traits::error_mapper::trait_def::ErrorMapper;
 
@@ -441,6 +445,30 @@ impl LLMProvider for OpenAIProvider {
         })
     }
 
+    async fn audio_transcription(
+        &self,
+        request: TranscriptionRequest,
+        _context: RequestContext,
+    ) -> Result<TranscriptionResponse, ProviderError> {
+        self.audio_transcription(request).await
+    }
+
+    async fn audio_translation(
+        &self,
+        request: TranslationRequest,
+        _context: RequestContext,
+    ) -> Result<TranslationResponse, ProviderError> {
+        self.audio_translation(request).await
+    }
+
+    async fn text_to_speech(
+        &self,
+        request: SpeechRequest,
+        _context: RequestContext,
+    ) -> Result<SpeechResponse, ProviderError> {
+        self.text_to_speech(request).await
+    }
+
     async fn health_check(&self) -> HealthStatus {
         let url = format!("{}/models?limit=1", self.config.get_api_base());
         let client = crate::core::http::outbound::default_outbound_client().clone();
@@ -677,108 +705,5 @@ impl LLMProvider for OpenAIProvider {
 }
 
 // Re-export error mapper from dedicated module
+pub use super::client_convenience::OpenAITask;
 pub use super::error_mapper::OpenAIErrorMapper;
-
-// ==================== Convenience Methods ====================
-
-impl OpenAIProvider {
-    /// Get model recommendations for specific use cases
-    pub fn get_recommended_model(&self, use_case: OpenAIUseCase) -> Option<String> {
-        get_openai_registry().get_recommended_model(use_case)
-    }
-
-    /// Check if a model supports a specific feature
-    pub fn model_supports_feature(&self, model_id: &str, feature: &OpenAIModelFeature) -> bool {
-        get_openai_registry().supports_feature(model_id, feature)
-    }
-
-    /// Get models by family (e.g., all GPT-4 variants)
-    pub fn get_models_by_family(&self, family: &super::models::OpenAIModelFamily) -> Vec<String> {
-        get_openai_registry().get_models_by_family(family)
-    }
-
-    /// Get models supporting specific feature
-    pub fn get_models_with_feature(&self, feature: &OpenAIModelFeature) -> Vec<String> {
-        get_openai_registry().get_models_with_feature(feature)
-    }
-
-    /// List available models from static registry
-    pub fn list_available_models(&self) -> Vec<String> {
-        self.models().iter().map(|m| m.id.clone()).collect()
-    }
-
-    /// Get model pricing information
-    pub fn get_model_pricing(&self, model_id: &str) -> Option<(f64, f64)> {
-        if let Ok(pricing) = crate::core::cost::calculator::get_model_pricing(model_id, "openai") {
-            return Some((
-                pricing.input_cost_per_1k_tokens,
-                pricing.output_cost_per_1k_tokens,
-            ));
-        }
-
-        if let Some(model_info) = self
-            .model_registry
-            .get_model_spec(model_id)
-            .map(|spec| &spec.model_info)
-            && let (Some(input_cost), Some(output_cost)) = (
-                model_info.input_cost_per_1k_tokens,
-                model_info.output_cost_per_1k_tokens,
-            )
-        {
-            return Some((input_cost, output_cost));
-        }
-        None
-    }
-
-    /// Get model context window information
-    pub fn get_model_context_window(&self, model_id: &str) -> Result<u32, ProviderError> {
-        let model_info = self.get_model_info(model_id)?;
-        Ok(model_info.max_context_length)
-    }
-
-    /// Check if model supports vision/multimodal input
-    pub fn model_supports_vision(&self, model_id: &str) -> bool {
-        self.model_supports_feature(model_id, &OpenAIModelFeature::VisionSupport)
-    }
-
-    /// Check if model supports function/tool calling
-    pub fn model_supports_tools(&self, model_id: &str) -> bool {
-        self.model_supports_feature(model_id, &OpenAIModelFeature::FunctionCalling)
-    }
-
-    /// Check if model supports streaming
-    pub fn model_supports_streaming(&self, model_id: &str) -> bool {
-        self.model_supports_feature(model_id, &OpenAIModelFeature::StreamingSupport)
-    }
-
-    /// Get best model for specific task
-    pub fn get_best_model_for_task(&self, task: OpenAITask) -> Option<String> {
-        match task {
-            OpenAITask::GeneralChat => self.get_recommended_model(OpenAIUseCase::GeneralChat),
-            OpenAITask::CodeGeneration => self.get_recommended_model(OpenAIUseCase::CodeGeneration),
-            OpenAITask::ComplexReasoning => self.get_recommended_model(OpenAIUseCase::Reasoning),
-            OpenAITask::VisionAnalysis => self.get_recommended_model(OpenAIUseCase::Vision),
-            OpenAITask::ImageGeneration => {
-                self.get_recommended_model(OpenAIUseCase::ImageGeneration)
-            }
-            OpenAITask::AudioTranscription => {
-                self.get_recommended_model(OpenAIUseCase::AudioTranscription)
-            }
-            OpenAITask::Embeddings => self.get_recommended_model(OpenAIUseCase::Embeddings),
-            OpenAITask::CostSensitive => self.get_recommended_model(OpenAIUseCase::CostOptimized),
-        }
-    }
-}
-
-/// OpenAI task categories for model selection
-#[derive(Debug, Clone)]
-pub enum OpenAITask {
-    GeneralChat,
-    CodeGeneration,
-    ComplexReasoning,
-    VisionAnalysis,
-    ImageGeneration,
-    AudioTranscription,
-    Embeddings,
-    CostSensitive,
-}
